third_party/blink/web_tests/platform/generic/http/tests/devtools/profiler/cpu-profiler-save-load-expected.txt

Tests that CPU profiling is able to save/load.

Profiler was enabled.

Running: testSave
PASS: open was called with extension '.cpuprofile'
PASS: write was called with data length more than zero
PASS: close was called

Running: testLoad
PASS: the content of loaded profile matches with the original profile.

Profiler was disabled.
